less leaks

+++ libft/srcs/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: aminko <aminko@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 16:58:02 by aminko            #+#    #+#             */
-/*   Updated: 2023/07/15 20:21:26 by aminko           ###   ########.fr       */
+/*   Updated: 2023/10/24 20:44:04 by aminko           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,15 +19,15 @@ caractères issue de la chaîne ’s’.
 Cette nouvelle chaîne commence à l’index ’start’ et
 a pour taille maximale ’len’. */
 
-char	*ft_substr(const char *s, unsigned int start, size_t len)
+char	*ft_substr(const char *s, int start, int len)
 {
-	size_t	i;
-	size_t	len_s;
+	int	i;
+	int	len_s;
 	char	*str;
 
 	if (!s)
 		return (NULL);
-	if (start >= (unsigned int)ft_strlen(s))
+	if (start >= ft_strlen(s))
 		return (str = ft_calloc(1, sizeof(char)));
 	i = 0;
 	len_s = ft_strlen(s);
